requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=5.4.0

# CLI enhancements
click>=8.0.0
colorama>=0.4.4
tqdm>=4.62.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Testing
pytest>=6.2.5
pytest-cov>=3.0.0
pytest-mock>=3.6.1

# Code quality
black>=21.9b0
flake8>=4.0.1
mypy>=0.910

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
